# backend/external-service/src/main/resources/api.yaml
swagger: '2.0'
info:
  description: Allu public interface. Allu is management system for Helsinki land areas.
  version: 1.0.0
  title: Allu public interface
  termsOfService: TODO
  contact:
    email: TODO@TODO.FI
  license:
    name: MIT
    url: 'https://bitbucket.org/vincit/allu/raw/HEAD/LICENSE'
securityDefinitions:
  api_key:
    type: apiKey
    name: Authorization
    in: header
basePath: /api-docs
tags:
  - name: applications
    description: Create and update applications
schemes:
  - http
paths:
  /v1/applications:
    post:
      tags:
        - applications
      summary: Create a new application
      description: Requires ROLE_INTERNAL or ROLE_TRUSTED_PARTNER
      operationId: applicationsCreate
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - in: body
          name: body
          description: Application created to Allu
          required: true
          schema:
            $ref: '#/definitions/ApplicationExt'
      responses:
        '200':
          description: Created application
          schema:
            $ref: '#/definitions/ApplicationExt'
      security:
        - api_key: []
  /v1/customers:
    post:
      tags:
        - customers
      summary: Create a new customer
      description: Requires ROLE_INTERNAL
      operationId: customersCreate
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - in: body
          name: body
          description: Customer created to Allu
          required: true
          schema:
            $ref: '#/definitions/CustomerExt'
      responses:
        '200':
          description: Created customer
          schema:
            $ref: '#/definitions/CustomerExt'
      security:
        - api_key: []
  /v1/customers/{id}:
    get:
      tags:
        - customers
      summary: Find a customer by id
      description: Requires ROLE_INTERNAL or ROLE_TRUSTED_PARTNER
      operationId: customersFindById
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - in: path
          name: id
          description: Id of the customer
          required: true
          type: string
      responses:
        '200':
          description: Customer with given id
          schema:
            $ref: '#/definitions/CustomerExt'
      security:
        - api_key: []
  /v1/contacts:
    post:
      tags:
        - contacts
      summary: Create a new contact
      description: Requires ROLE_INTERNAL or ROLE_TRUSTED_PARTNER
      operationId: contactsCreate
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - in: body
          name: body
          description: Contact created to Allu
          required: true
          schema:
            $ref: '#/definitions/ContactExt'
      responses:
        '200':
          description: Created contact
          schema:
            $ref: '#/definitions/ContactExt'
      security:
        - api_key: []
  /v1/contacts/{id}:
    get:
      tags:
        - contacts
      summary: Find a contact by id
      description: Requires ROLE_INTERNAL or ROLE_TRUSTED_PARTNER
      operationId: contactsFindById
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - in: path
          name: id
          description: Id of the contact
          required: true
          type: string
      responses:
        '200':
          description: Contact with given id
          schema:
            $ref: '#/definitions/ContactExt'
      security:
        - api_key: []
definitions:
  CustomerType:
    description: Type of a customer
    type: string
    enum:
      - PERSON
      - COMPANY
      - ASSOCIATION
      - PROPERTY
      - OTHER
  StatusType:
    description: Application status
    type: string
    enum:
      - PRE_RESERVED
      - PENDING
      - HANDLING
      - RETURNED_TO_PREPARATION
      - DECISIONMAKING
      - DECISION
      - REJECTED
      - FINISHED
      - CANCELLED
  ApplicationType:
    description: Application type
    type: string
    enum:
      - EXCAVATION_ANNOUNCEMENT
      - AREA_RENTAL
      - TEMPORARY_TRAFFIC_ARRANGEMENTS
      - CABLE_REPORT
      - PLACEMENT_CONTRACT
      - EVENT
      - SHORT_TERM_RENTAL
      - NOTE
  ApplicationKind:
    description: >-
      Application type specifies the subtype of the application. Kind is
      application type specific and all kinds are not allowed for all types (see
      Java code for exact information)
    type: string
    enum:
      - PROMOTION
      - OUTDOOREVENT
      - BRIDGE_BANNER
      - BENJI
      - PROMOTION_OR_SALES
      - URBAN_FARMING
      - KESKUSKATU_SALES
      - SUMMER_THEATER
      - DOG_TRAINING_FIELD
      - DOG_TRAINING_EVENT
      - SMALL_ART_AND_CULTURE
      - SEASON_SALE
      - CIRCUS
      - ART
      - STORAGE_AREA
      - STREET_AND_GREEN
      - WATER_AND_SEWAGE
      - ELECTRICITY
      - DATA_TRANSFER
      - HEATING_COOLING
      - CONSTRUCTION
      - YARD
      - GEOLOGICAL_SURVEY
      - PROPERTY_RENOVATION
      - CONTAINER_BARRACK
      - PHOTO_SHOOTING
      - SNOW_WORK
      - RELOCATION
      - LIFTING
      - NEW_BUILDING_CONSTRUCTION
      - ROLL_OFF
      - CHRISTMAS_TREE_SALES_AREA
      - CITY_CYCLING_AREA
      - AGILE_KIOSK_AREA
      - STATEMENT
      - SNOW_HEAP_AREA
      - SNOW_GATHER_AREA
      - OTHER_SUBVISION_OF_STATE_AREA
      - MILITARY_EXCERCISE
      - WINTER_PARKING
      - REPAVING
      - ELECTION_ADD_STAND
      - PUBLIC_EVENT
      - OTHER
  ApplicationTagType:
    description: Application tag type
    type: string
    enum:
      - ADDITIONAL_INFORMATION_REQUESTED
      - STATEMENT_REQUESTED
      - DEPOSIT_REQUESTED
      - DEPOSIT_PAID
      - PRELIMINARY_INSPECTION_REQUESTED
      - PRELIMINARY_INSPECTION_DONE
      - FINAL_INSPECTION_AGREED
      - FINAL_INSPECTION_DONE
      - WAITING
      - COMPENSATION_CLARIFICATION
      - PAYMENT_BASIS_CORRECTION
  CustomerRoleType:
    description: Application customer role
    type: string
    enum:
      - APPLICANT
      - PROPERTY_DEVELOPER
      - CONTRACTOR
      - REPRESENTATIVE
  TrafficArrangementImpedimentType:
    description: Type of the traffic impediment
    type: string
    enum:
      - NO_IMPEDIMENT
      - SIGNIFICANT_IMPEDIMENT
      - IMPEDIMENT_FOR_HEAVY_TRAFFIC
      - INSIGNIFICANT_IMPEDIMENT
  CustomerExt:
    description: Customer information
    type: object
    required:
      - type
      - name
    properties:
      id:
        description: The unique identifier of the customer
        type: integer
        format: int32
      type:
        $ref: '#/definitions/CustomerType'
      name:
        description: Name of the customer
        type: string
      postalAddress:
        $ref: '#/definitions/PostalAddressExt'
      email:
        description: Email of the customer
        type: string
      phone:
        description: Phone number of the customer
        type: string
      registryKey:
        description: Registry key of the customer. Either business id (Y-tunnus) for companies or social security number (henkilötunnus) for persons
        type: string
  ContactExt:
    description: Customer contact information
    type: object
    properties:
      id:
        description: The unique identifier of the contact
        type: integer
        format: int32
      customerId:
        description: The unique identifier of the customer
        type: integer
        format: int32
      name:
        description: Name of the contact
        type: string
      postalAddress:
        $ref: '#/definitions/PostalAddressExt'
      email:
        description: Email of the customer
        type: string
      phone:
        description: Phone number of the customer
        type: string
  ApplicationExt:
    type: object
    required:
      - status
      - type
      - kind
      - name
      - creationTime
      - extension
    properties:
      id:
        description: The unique identifier of the application
        type: integer
        format: int32
      projectId:
        description: The unique identifier of the project the application belongs to
        type: integer
        format: int32
      customersWithContacts:
        type: array
        items:
          $ref: '#/definitions/CustomerWithContactsExt'
      locations:
        type: array
        items:
          $ref: '#/definitions/LocationExt'
      status:
        $ref: '#/definitions/StatusType'
      type:
        $ref: '#/definitions/ApplicationType'
      kind:
        $ref: '#/definitions/ApplicationKind'
      applicationTags:
        type: array
        items:
          $ref: '#/definitions/ApplicationTagExt'
      name:
        description: Name of the application
        type: string
      creationTime:
        description: Creation time of the application
        type: string
        format: date-time
      startTime:
        description: >-
          start time of the application. This is automatically calculated from
          location data
        type: string
        format: date-time
      endTime:
        description: >-
          end time of the application. This is automatically calculated from
          location data
        type: string
        format: date-time
      extension:
        $ref: '#/definitions/ApplicationExtensionExt'
  ApplicationTagExt:
    type: object
    description: Tag of an application
    properties:
      type:
        $ref: '#/definitions/ApplicationTagType'
      creationTime:
        description: Creation time of the application tag
        type: string
        format: date-time
  CustomerWithContactsExt:
    type: object
    properties:
      roleType:
        $ref: '#/definitions/CustomerRoleType'
      customer:
        description: Id of the customer
        type: integer
        format: int32
      contacts:
        description: Contacts of the customer
        type: array
        items:
          type: integer
          format: int32
  Geometry:
    description: Any multipolygon GeoJSON
    type: object
  LocationExt:
    description: Description of time and location of an application
    type: object
    properties:
      id:
        description: Id of the location
        type: integer
        format: int32
      locationKey:
        description: >-
          If an application has multiple locations (applies to limited
          application types), location key contains the number of location
        type: integer
        format: int32
      locationVersion:
        description: Version of the location
        type: integer
        format: int32
      startTime:
        description: Start time of the use of location
        type: string
        format: date-time
      endTime:
        description: End time of the use of location
        type: string
        format: date-time
      additionalInfo:
        description: Additional info about the location
        type: string
      geometry:
        $ref: '#/definitions/Geometry'
      area:
        description: Area in square meters calculated from the geometry
        type: number
        format: double
      areaOverride:
        description: >-
          Area in square meters as defined manually. If defined, calculated area
          is not used
        type: number
        format: double
      postalAddress:
        $ref: '#/definitions/PostalAddressExt'
      paymentTariff:
        description: Payment tariff calculated from the location geometry
        type: integer
        format: int32
      paymentTariffOverride:
        description: >-
          Payment tariff set manually. If defined, calculated payment tariff is
          not used
        type: integer
        format: int32
      underpass:
        description: 'True, if it''s possible to pass the location below construction area'
        type: boolean
  PostalAddressExt:
    description: Postal address
    type: object
    properties:
      streetAddress:
        description: Street address. For example Mannerheimintie 1
        type: string
      postalCode:
        description: Postal code of address. For example 00100
        type: string
      city:
        description: City of postal address. For example Helsinki
        type: string
  ApplicationExtensionExt:
    description: >-
      Base class for application extensions. Application extension contains
      application type specific data. See swagger.json and types referring to
      this type using allOf : [{$ref : #/definitions/ApplicationExtensionExt}]
    type: object
    discriminator: applicationType
    properties:
      terms:
        type: string
      applicationType:
        $ref: '#/definitions/ApplicationType'
  ExcavationAnnouncementExt:
    description: Extends ApplicationExtensionExt. Excavation announcement data.
    type: object
    allOf:
      - $ref: '#/definitions/ApplicationExtensionExt'
      - type: object
        properties:
          pksCard:
            description: >-
              PKS-kortti (pääkaupunkiseudun katutöihin liittyvät koulutus).
              True, if the contractor doing the actual work has PKS-card.
            type: boolean
          constructionWork:
            description: Construction
            type: boolean
          maintenanceWork:
            description: Maintenance
            type: boolean
          emergencyWork:
            description: Emergency work
            type: boolean
          propertyConnectivity:
            description: Property connectivity (kiinteistöliitos)
            type: boolean
          winterTimeOperation:
            description: >-
              Date of winter time operability (Talvityön toiminnallinen kunto,
              valmistumis pvm)
            type: string
            format: date-time
          workFinished:
            description: The date work has been accepted as finished
            type: string
            format: date-time
          unauthorizedWorkStartTime:
            description: The time un-authorized work began
            type: string
            format: date-time
          unauthorizedWorkEndTime:
            description: The time un-authorized work ended
            type: string
            format: date-time
          guaranteeEndTime:
            description: Date of the guarantee end time
            type: string
            format: date-time
          cableReportId:
            description: Id of the cable report related to this excavation announcement
            type: integer
            format: int32
          additionalInfo:
            description: Additional information regarding application
            type: string
          trafficArrangements:
            description: Information regarding traffic arrangements
            type: string
          trafficArrangementImpedimentType:
            $ref: '#/definitions/TrafficArrangementImpedimentType'